+++ paper.bib
@@ -56,7 +56,7 @@
     publisher = {Zenodo},
     year = {2019},
     version = {v2019.06.1},
-    doi = {https://doi.org/10.5281/zenodo.3266823},
+    doi = {10.5281/zenodo.3266823},
 }
 
 @misc{hpc_shell,
@@ -77,7 +77,7 @@
     author = {Cathrine Bergh and Maryam Majdolhosseini and Alessandra Villa},
     title = {GROMACS tutorial: Introduction to Molecular Dynamics},
     year = {2025},
-    doi = {https://doi.org/10.5281/zenodo.14803238}},
+    doi = {10.5281/zenodo.14803238}},
 }
 
 @article{vmd,
@@ -87,6 +87,7 @@
   year={1996},
   volume={14},
   pages={33-38},
+  doi={10.1016/0263-7855(96)00018-5}
 }
 
 @article{gromacs,
@@ -96,7 +97,7 @@ volume = {1-2},
 pages = {19-25},
 year = {2015},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2015.06.001},
+doi = {10.1016/j.softx.2015.06.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
 author = {Mark James Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
 }
@@ -116,6 +117,6 @@ volume = {8},
 issue = {83},
 pages = {265},
 year = {2025},
-doi = {https://doi.org/10.21105/jose.00265},
+doi = {10.21105/jose.00265},
 author = {Matteo T. Degiacomi, Richard J. Gowers, Micaela Matta, and Antonia S. J. S. Mey},
 }
