### 2025-06-25 85c0909
adding JOSE badge
