### 2025-06-25 #287 @arm61
@sarahvs99, looks good. I note that there are two authors on the paper that haven’t contributed to the repository (Hannah Pollak and Angela Chitzanidi). Can you comment on their contribution?
