### 2025-06-25 #287 @sarahvs99
@arm61 The readme.md has a complete overview of what everyone did. In short, Hannah wrote a lot of the initial version of the course, with all of her commits on a private repo that we use. Hannah also helped massively with the setup of the software and the virtual machines. Angela was in charge of the computational resources (software, HPC, virtual machines) for both years the course has run and advised a lot on what we could and couldn't do. She wrote the section of the paper on the resources, but sent it to me privately instead of on GitHub.
