### 2025-06-25 #287 @ashleefv
A few other reviewer suggestions: Jindal Shah (ShahResearchGroup), Eric Jankowski (erjank), Michael Shirts (mrshirts), Sapna Saurpria (SarupriaGroup), or others who have been contributors to the iCoMSE educational workshop community. Team | i-CoMSE https://share.google/TT8MBckHU989fwOkR
