### 2025-07-24 #287 @m-rivera
Hi folks,

Sorry, I haven't reviewed JOSE before. Should I wait for the second reviewer to be assigned before going through the [checklist](https://openjournals.readthedocs.io/en/jose/review_checklist.html)?

I've only spotted a few minor things to suggest, otherwise the repo and paper are in excellent shape in my opinion.
