### 2025-07-24 #299 @editorialbot
[REVIEW]: Course Materials for an Introduction to Computational Chemistry Techniques
**Submitting author:** @sarahvs99 (<a href="http://orcid.org/0000-0003-1663-5206">Sarah Stewart</a>)
**Repository:** https://github.com/Edinburgh-Chemistry-Teaching/MD_ResearchTechniques
**Branch with paper.md** (empty if default branch):
**Version:** 1.1
