### 2025-07-24 #299 @arm61
@m-rivera and @ambrown77, thanks again for agreeing to review. If you work through the checklist (information above on how to generate it) and if there are any problems/comments about the material, I would recommend opening an issue on [the material repository](https://github.com/Edinburgh-Chemistry-Teaching/MD_ResearchTechniques), and the authors can sort them out.

More information about the review guidelines can be found on the Open Journals documentation pages: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html

If anyone has any questions just ping me on here!
