### 2025-07-24 Edinburgh-Chemistry-Teaching/MD_ResearchTechniques#3 @m-rivera
JOSE Review
First of all, I wanted to commend your work on making this resource and on making it publicly accessible. From experience, I've rarely seen so much depth in bespoke teaching material for scientific computing at PGT level. The commitment to teaching so many useful command line skills is commendable, given how easy it is to allow students to work in inefficient ways and avoid the subjects of loops, pipes etc. altogether (or to defer to general purpose tutorials without the targeted scope of MD simulation). The quality of figures is especially excellent and should help in teaching such technical concepts.

It's also a good idea to have shared the teaching material in a portable format, although as you will see below, I think a bit of clarification would help someone else make the best use of your html pages.

Here are the parts of the checklist which I think need addressing:

> I didn't see a version number.

> The dependencies are in the paper but would be welcome in the README as well.
> It would be nice to flag areas of the notes which will need editing if using your lessons outside of Edinburgh. For example, you could warn the user to replace the Eddie instructions in Session 1 Part 3 with their own teaching cluster. You could also mention the parts of Session 4 which would need filling out (as you state in the paper).
> I couldn't find these.

> A bit more would be welcome here. In particular, it wasn't very clear where the `Notes/` were hosted and whether you imagine someone else putting them in their VLE, an independent website, or just delivered to students for them to open with a browser locally.

### Other
Optional things to address:
- [ ] The paper makes reference to 'feedback from students' but it's unclear how this feedback was gathered or whether ethical approval was granted to share it. It could be safer to share instructor's anecdotal impressions, which are still useful but raise fewer questions.
- [ ] You explain the assessment is made up of quizzes and project work. It would be useful to know the setting of the quizzes, specifically if they were timed and/or invigilated.
- [ ] Session 3 has links to an open Wooclap at the bottom of the TOC page. You might want to delete that, or replace with a placeholder link.

Thanks again for writing this up!
