### 2025-08-07 94fd59c
Update README.md
Included a section for gromacs version and VMD for rendering
