### 2025-09-08 36b8846
Update README.md with clearer instructions on how to contribute
