+++ README.md
@@ -48,16 +48,41 @@ Please note that we are not sharing the materials for Session 4, as those are us
 - [ClayCode workshop](https://github.com/Erastova-group/ClayCode-workshop)
 - [VMD](https://www.ks.uiuc.edu/Research/vmd/)
 
+
 ## Gromacs version and others:
 - Gromacs  version, `gromacs/2024.4`,  was used for simulations throughout this course.
 - VMD was used for rendering of images.
 
-## Reuse and adaptation of the material
 
-Contributions are welcome and can be made through creating an issue or a [pull request](https://docs.github.com/en/get-started/quickstart/contributing-to-projects).
+## Contributing to the resource and raising issues
 
-To reuse this material, you can [clone](https://docs.github.com/en/repositories/creating-and-managing-repositories/cloning-a-repository) this repository and adapt it to your needs.
+Contributions to the learning resource are welcome. Contributions can be made through creating an issue or a pull request.
+
+### For issues
+- To create an issue, contributors are encouraged to follow the [GitHub quickstart guide on creating an issue.](https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue)
+- Make sure to include the following into your issue:
+	-  What version of the different packages are you running
+	-  Are you using it as a student or instructor?
+	-  Is the issue reporting a bug, an enhancement, or a feature request		
+
+### For pull requests
+- To create a pull request, contributors are encouraged to follow the [GitHub quickstart guide on creating a fork and submitting a pull request.](https://docs.github.com/en/get-started/quickstart/contributing-to-projects)
+
+### Telling us about how you are using the resource
+If you just want to tell us how you have been using the resource just send us an email or raise an issue pointing to your work. 
+
+## Reusing and building on the material for your own course
+
+The easiest way is by cloning the material and adapting it to your needs. This can be just using some partial material or expanding on the existing material. The best way to do this is by either [cloning](https://docs.github.com/en/repositories/creating-and-managing-repositories/cloning-a-repository) the repo and building up on it, or using the current repository as a [template repository](https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-repository-from-a-template) for your own or your organisations GitHub account, for more details see the [overview](overview.md) document. 
+
+
+## Reference:
 
 If using this material, please acknowledge the developers and cite the associated publication:
 
 - JOSE REFERENCE WHEN AVAILABLE
+
+
+## License
+
+These materials are made freely available, and are licensed under a [CC-BY 4.0](https://creativecommons.org/licenses/by/4.0/) license.
