### 2025-09-08 8301aef
Update paper.md summary
