### 2025-09-08 b448c2e
Update README.md on reuse of resources
