### 2025-09-18 Edinburgh-Chemistry-Teaching/MD_ResearchTechniques#3 @m-rivera
Thanks very much for addressing these comments. I want to push back one last time against including research data on human participants, even statistical, without having sought ethical approval before gathering the data.

I realise that ethical approval processes are often inappropriate for Chemistry departments, indeed the only references I could find at the University of Edinburgh were [from](https://health.ed.ac.uk/research/ethics-and-integrity) [other](https://ppls.ed.ac.uk/psychology/research/research-ethics) [departments](https://llc.ed.ac.uk/research/ethics). Nevertheless, I gathered that the institution adheres to the [UK Research Integrity Office Code of Practice for Research](https://ukrio.org/wp-content/uploads/UKRIO-Code-of-Practice-for-Research.pdf).

Sections 3.6.5 and 3.6.8 place the responsibility on the institution and the researcher to run these research projects by an ethical approval process.

I realise that it's a grey area whether this constitutes _research_, but I think it would go a long way to protecting your liability if you removed the numbers and instead stated anecdotal impressions. Yes, it would weaken your statements, but if you rely on the numbers to make a strong logical inference in your paper, to me it gets dangerously close to "research" territory.

As I said in the first place, I consider this an optional change, as I couldn't see that it was an explicit requirement for JOSE review. If you wish to go ahead anyway, I will recommend the paper for approval.
