### 2025-09-18 #299 @m-rivera
I'm close to it, just want a final bit of clarification about ethical approval:
https://github.com/Edinburgh-Chemistry-Teaching/MD_ResearchTechniques/issues/3#issuecomment-3306021838
