+++ paper.md
@@ -193,8 +193,7 @@ com/Erastova-group/ClayCode-workshop) introducing the use of ClayCode[@pollak202
 
 # Acknowledgements
 We thank the demonstrators, Finlay Clark, Chenfeng Zhang and Ryan Zhu, who supported student learning and Rosa Brauer for feedback on the instructional materials.
-We also thank Mike Wallis for assistance with setting up the original computational resources, David August for organising the course this module was part of, and Matteo Degiacomi for providing protein lecture material for session 3.
-This work builds on contributions from Software Carpentry and GROMACS community tutorials. 
+We also thank Mike Wallis for assistance with setting up the original computational resources, David August for organising the course this module was part of, and Matteo Degiacomi for providing protein lecture material for session 3. This work builds on contributions from Software Carpentry and GROMACS community tutorials. Student feedback mentioned here is based on informal feedback collected as part of routine course delivery and not as part of a formal research study.
 
 
 
