### 2025-09-18 Edinburgh-Chemistry-Teaching/MD_ResearchTechniques#3 @punkpony
Thank you @m-rivera for the info re ethical approval. The feedback gathered was informal since we were rolling out the module,   and we do not use it for research study purposes (I believe there was a formal one run by the university, but we do not use it). We do not present any numerical data on the feedback, and so we did not think there was a need for ethical approval. We will keep this in mind for future works we do:)
To make sure this is clear in the paper, we added “Student feedback mentioned here is based on informal feedback collected as part of routine course delivery and not as part of a formal research study" to the Acknowledgements section. Do you think this is sufficient?
