### 2025-09-18 Edinburgh-Chemistry-Teaching/MD_ResearchTechniques#3 @m-rivera
Yes, I think that's now completely safe. Thanks for addressing it, I know it can be irritating when it feels like no one is at any remote risk of harm.
