### 2025-09-18 #299 @m-rivera
Excellent! I've updated the checklist and am happy to recommend accepting the work.

Let me know if there's anything else to do!
