### 2025-09-18 #299 @ambrown77
I have uploaded one final comment related to learning objectives that I think is still important for clarity for others using this content. I have uploaded the checklist to reflect the rest of the response to comments.
