### 2025-09-25 f205e6b
Update paper.md
changing Learning Objectives section title to Module Goals
