### 2025-09-25 #299 @punkpony
We changed the sub-section title from 'Learning Objectives' to 'Module Goals'. I note that the JOSE checklist does not require an explicit list of LO.
