### 2025-09-25 #299 @arm61
I think that sounds like a good compromise! @ambrown77 I feel that should be sufficient to recommend accepting this work?
