### 2025-09-25 #299 @punkpony
Thank you @arm61, @ambrown77 and @m-rivera for your time and valuable input - we look forward to seeing the paper published in JOSE!
