### 2025-09-26 c564cd2
Fixing author names paper.bib
