+++ paper.bib
@@ -118,5 +118,5 @@ issue = {83},
 pages = {265},
 year = {2025},
 doi = {10.21105/jose.00265},
-author = {Matteo T. Degiacomi, Richard J. Gowers, Micaela Matta, and Antonia S. J. S. Mey},
+author = {Matteo T. Degiacomi and Richard J. Gowers and Micaela Matta, and Antonia S. J. S. Mey},
 }
