### 2025-09-26 fe9080c
Updating paper.md
