+++ paper.md
@@ -188,7 +188,7 @@ This module equips postgraduate students with practical computational chemistry 
 It is particularly suited to courses seeking to integrate practical computational methods skills with experimental and theoretical chemistry teaching.
 
 For those interested in learning more about simulating clay surfaces, we have also developed a [workshop](https://github.
-com/Erastova-group/ClayCode-workshop) introducing the use of ClayCode[@pollak2024] software for set-up and simulation of realistic clay mineral systems.
+com/Erastova-group/ClayCode-workshop) introducing the use of ClayCode [@pollak2024] software for set-up and simulation of realistic clay mineral systems.
 
 
 # Acknowledgements
