### 2025-09-26 #299 @sarahvs99
@arm61 for one of the carpentry references, the author is just "HPC Carpentry" as the author list on the GitHub is empty. Should I leave it as this, or use the list of contributors (with IRL names instead of username if I can find them)?
