### 2025-09-26 #299 @arm61
Do the carpentries have guidance on referencing the materials?
