### 2025-09-26 2fd96d0
Update paper.bib
