+++ paper.bib
@@ -1,5 +1,5 @@
 @misc{ecdf,
-    author = {University of Edinburgh},
+    author = {{University of Edinburgh}},
     title = {Edinburgh Compute and Data Facility},
     howpublished = {\url{https://www.ecdf.ed.ac.uk}},
      note = {Accessed April 16, 2025}
@@ -60,7 +60,7 @@
 }
 
 @misc{hpc_shell,
-    author = {HPC Carpentry},
+    author = {{HPC Carpentry development team}},
     title = {hpc-carpentry/hpc-shell: Introduction to Using the Shell in a High-Performance Computing Context},
     year = {2021},
     url = {https://hpc-carpentry.github.io/hpc-shell/},
