### 2025-09-26 #299 @punkpony
@sarahvs99 @arm61 I agree, keep as HPC Carpentry, sometimes communities use their name as an author in papers, see OOLEN as example.
