### 2025-09-26 #299 @sarahvs99
@arm61 we changed a few things and everything looks as it should as far as we can see. The only thing is that the "archive" link on the front page currently links to doi.org. Is there anything we need to do to fix that?
