### 2025-09-29 #299 @arm61
Hey @sarahvs99, that is fine. That will be updated when the paper is properly published.

I have looked through the text and I am all happy. The next step is for you to make a Zenodo archive of the repository and let me know the DOI of the archive. Please also provide the version number of the archived repository.
