### 2025-09-29 #299 @sarahvs99
The doi is https://doi.org/10.5281/zenodo.17226164, and the version is 1.1
