### 2025-09-29 #299 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
