### 2025-11-18 #299 @labarba
Congratulations @sarahvs99 — your JOSE paper is now published! 🚀

Tremendous thanks to our Editor: @arm61 and Reviewers: @m-rivera, @ambrown77 — we couldn't do this without you 🙏
